All structs are now in file type.jl
The stucts are now under the unbrella xyTypes. The purpose of this is
that when a struct is change, there is only a warning message. There
is no need to restart Julia, or recompile the package.
