+++ Sync.sh
@@ -3,8 +3,11 @@
 if [ $# -eq 0 ]; then # copy codes to remote
     rsync -avh --exclude rst --exclude ".*" . amd:/mnt/a/store/xybng/a
     rsync -avh --exclude rst --exclude ".*" . amd:/mnt/a/store/xybng/b
-elif [ $# -eq 2 ]; then # copy remote summary results to local
-    scp amd:/mnt/a/store/xybng/a/rst/$1/s* rst/$2
+elif [ $# -eq 3 ]; then # copy remote summary results to local
+    if [ ! -d rst/$3 ]; then
+        mkdir -p rst/$3
+    fi
+    scp amd:/mnt/a/store/xybng/$1/rst/$2/s* rst/$3
 else
     echo Not defined
 fi
