+++ src/xy.jl
@@ -24,128 +24,7 @@ module XY
 using xyBnG.Util
 using Mmap
 using Serialization
-
-const _nvldtype = 13
-
-"""
-    _type(x::Union{Int8, DataType})
-
-A literal bidirectional mapping between `Int8` and `DataType`.
-This is used to decide the element type of the matrix,
-or the type tag in the header of an `xy` file.
-
-No range test, as this is an internal function. The valid types are
-`Int8`, `Int16`, `Int32`, `Int64`, `Int128`, `UInt8`, `UInt16`, `UInt32`,
-`UInt64`, `UInt128`, `Float32`, and `Float64`.
-"""
-function _type(x::Union{Int8,DataType})
-    vt = (
-        Int8,
-        Int16,
-        Int32,
-        Int64,
-        Int128,
-        UInt8,
-        UInt16,
-        UInt32,
-        UInt64,
-        UInt128,
-        Float32,
-        Float64,
-        Bool,
-    )
-    if x isa Int8
-        vt[x]
-    else
-        Int8(findfirst(t -> t == x, vt))
-    end
-end
-
-"""
-    mutable struct header
-
-The header struct of an `xy` file. It consists of 8 `Int8` fields.
-- x::Int8
-- y::Int8
-- v::Int8      # x, y, v should be 'x', 'y', ' '. they are the magic chars
-- flus::Int8   # FLUS matrix type: full, lower triangle, upper triangle, symmetric(lower)
-- major::Int8  # 0 for loci majored, 1 for ID majored, or else
-- type::Int8   # element type of the matrix, determined by function _type
-- r::Int8      # 1 for BitArray. 0 for others
-- u::Int8      # r and u are reserved
-It is by default as (x = 'x', y = 'y', v = ' ', flus = 'F', major = 0, type = 1, r = 0, u = 0).
-Or one can specify the fields by keyword arguments. For example,
-```julia
-header(x = 'x', y = 'y', v = ' ', flus = 'F', major = 0, type = Int8, r = 0, u = 0)
-```
-
-## Specification of `u`
-- 0 for SNP coding
-- 1 for IBD coding
-- 2 for genotype coding
-- 3+ else
-"""
-mutable struct header
-    x::Int8
-    y::Int8
-    v::Int8      # x, y, v should be 'x', 'y', ' '. they are the magic chars
-    flus::Int8   # FLUS matrix type: full, lower triangle, upper triangle, symmetric
-    major::Int8  # 0 for loci majored, 1 for ID majored, or else
-    type::Int8   # element type of the matrix, determined by function _type
-    r::Int8      # 1 for BitArray. 0 for others
-    u::Int8      # 0 for SNP coding, 1 for IBD coding, 2 for genotype coding, 3+ else
-    function header(;
-        flus = 'F',  # full, lower triangle, upper triangle, symmetric
-        major = 0,
-        type = 1,
-        u = 0,
-    )
-        flus ∉ "FLUS" ||
-            major ∉ 0:1 ||
-            type ∉ 1:_nvldtype ||
-            u < 0 && error("Invalid header")
-        new('x', 'y', ' ', flus, major, type, 0, u)
-    end
-end
-
-function Base.show(io::IO, h::header)
-    if h.r == 1
-        print(io, "A BitArray matrix")
-    else
-        print(io, "A normal matrix")
-    end
-    if h.u == 0
-        println(io, " coding SNP alleles.")
-    elseif h.u == 1
-        println(io, " coding IBD alleles.")
-    elseif h.u == 2
-        println(io, " coding genotypes.")
-    else
-        println(io, " with unknown coding.")
-    end
-    println(io, "        -------")
-    m = join(Char.((h.x, h.y, h.v)))
-    println(io, "The magic chars: '$m'.")
-    if h.flus == Int8('F')
-        println(io, "         Matrix: full matrix.")
-    elseif h.flus == Int8('L')
-        println(io, "         Matrix: lower triangle of a square matrix.")
-    elseif h.flus == Int8('U')
-        println(io, "         Matrix: upper triangle of a square matrix.")
-    elseif h.flus == Int8('S')
-        println(io, "         Matrix: symmetric matrix.")
-    else
-        println(io, "         Matrix: unknown.")
-    end
-    if h.major == 0
-        println(io, "          Major: loci majored.")
-    elseif h.major == 1
-        println(io, "          Major: ID majored.")
-    else
-        println(io, "          Major: $(h.major) -- unknown.")
-    end
-    println(io, "   Element type: $(_type(h.type))")
-end
+import xyBnG.xyTypes: header, _type, _nvldtype
 
 """
     isEqual(h1::T, h2::T) where T <: header
